+++ example/package.json
@@ -22,6 +22,11 @@
   "eslintConfig": {
     "extends": "react-app"
   },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
+  },
   "browserslist": [
     ">0.2%",
     "not dead",
