+++ package.json
@@ -81,5 +81,10 @@
         "lines": 90
       }
     }
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
   }
 }
